feat: add snackbar on edit permission
